KVM test: Change variant name from virtio to virtio_net
virtio as the variant name for virtio nw interface
tests can be confusing, so let's try to be consistent
with the virtio block variant (virtio_blk), and change
virtio to virtio_net. Since Fedora 13 already supports
virtio out of the box, let's make it the default on the
default test set qemu_kvm_f13_quick.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4771 592f7852-d20e-0410-864c-8624ca9c26a4