KVM test: virtio_console test v3
1) Starts VMs with the specified number of virtio console devices
2) Start smoke test
3) Start loopback test
4) Start performance test

This test uses an auxiliary script, console_switch.py, that is
copied to guests. This script has functions to send and write
data to virtio console ports. Details of each test can be found
on the docstrings for the test_* functions.

Changes from v2:
 * Fix a bug that prevented the virtio_console auxiliary script
   to work under Fedora 14

Changes from v1:
 * Style fixes
 * Whitespace cleanup
 * Docstring and message fixes
 * ID for char devices can't be a simple number, fix it

Tested with Fedora 13 guest/host, Fedora 14 guest -smp 1 and
-smp 2 (smp 2 triggered some bugs on the guest kernel)

Signed-off-by: Lukas Doktor <ldoktor@redhat.com>
Signed-off-by: Jiri Zupka <jzupka@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4772 592f7852-d20e-0410-864c-8624ca9c26a4