KVM test: tests.cfg.sample: add params to qemu command line properly
There's a small bug on the qemu sample testset for KVM autotest:
We need to append -enable-kvm as extra params to the qemu command
line, not replace all extra params. Small bug, renders the unattended
install unusable for folks trying qemu instead of qemu-kvm.

Thanks to Kevin Wolf for having found this problem.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4778 592f7852-d20e-0410-864c-8624ca9c26a4