server/git.py: Simplify methods get_local_head and get_remote_head
It is not necessary to resort to the git web frontend
to verify whether the local repo is outdated with regards
to origin. This patch simplifies the methods using built
in git commands, thus eliminating the need of a web URL.

The interface was maintainted, and now people who
still specify a web URL on their control file will have
a deprecationwarning message on their logs.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4781 592f7852-d20e-0410-864c-8624ca9c26a4