Fixing ftrace profiler from r4787
David Sharp <dhsharp@google.com> has noticed that
some attributes attribution were being done on the
setup() stage, that gets executed only if version
changes, so the profiler would fail on a second
execution of an autotest job on a particular machine.

So move the attributions to the initialize() method,
and the trace-cmd reset command to the start()
method.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@4788 592f7852-d20e-0410-864c-8624ca9c26a4