KVM test: Fix command to install virtio driver msi install package
If we are going to use the msi package, we have to tell windows
to install it using msiexec /passive /package, so the msi is
installed without asking questions. Only executing the msi
will prompt questions to the user, which we clearly don't want.

Changing it on the default config file so people won't loose
time trying to figure out why the msi is not being installed.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4791 592f7852-d20e-0410-864c-8624ca9c26a4