KVM-test: Update pci_hotplug to suit for new qemu
In the new version of qemu device_add is instead of pci_add. So modify
the scripts to suit for the new version of qemu. Now both nic and block
devices are support in qemu, so keep using monitor command line for
this test.

Changs from v1:
Can support block device hotplug in scripts. But scsi is still not
supported.

Changs from v2:
 - using "?" to check which command is supported in qemu and add the
   debug information for unsupported commands.
 - using "$add_cmd ?" to check the support devices in command line and
   add the debug information for unsupported devices

Signed-off-by: Yiqiao Pu <ypu@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4808 592f7852-d20e-0410-864c-8624ca9c26a4