Fix partition.mount() to use the default mount point
If you setup a partition object with a mount point, and then call
partition.mount(), python will complain because mountpoint is a required option,
but thats not the case if you specified a mount when you created the object.  So
set mountpoint=None so that we can properly use the mountpoint we were created
with.

Signed-off-by: Josef Bacik <josef@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4813 592f7852-d20e-0410-864c-8624ca9c26a4