client/bin/kernel.py: default kernel arguments shall be kept
When adding a new kernel to bootload we want the arguments of
the default entry to be kept in the new entry, otherwise that
entry might not boot (example, if we disregard LVM parameters
and the root is on an LVM volume, the entry won't boot).

So, instead of setting the arguments while the kernel entry
is being created (doing this with boottool will completely
disregard any arguments of the default entry), create the
new entry *then* add the additional arguments, one by one.
In every new addition, if the arg is already on the arg list,
the arg won't be added, therefore avoiding duplicate args.

Changes from v1:
- Fix kernel_unittest.py

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4851 592f7852-d20e-0410-864c-8624ca9c26a4