QMPMonitor: Introduce the get_greeting() method
It returns the QMP's greeting message as sent by the monitor.

Please, note that this commit also changes the QMPMonitor's constructor
to store the full greeting message (it currently stores only its
contents).

This new method is going to be used by the QMP test-suite, which fully
checks the greeting message.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4868 592f7852-d20e-0410-864c-8624ca9c26a4