tiobench test: Actually report results
If you want to find the results from tiobench, you currently have to dig
through debug logs.  This should be an improvement on that.

Changes from v1:
* Do the results reporting similarly to what's been done in other
benchmarks.
* Instead of doing the default 2 runs at one autotest iteration,
slightly modify the default control file to run 2 iterations
using the autotest engine, and exemplifying the use of the
iterations argument that job.run_test() can take.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4877 592f7852-d20e-0410-864c-8624ca9c26a4