Autotest client sample control files update
The samples are one of the first contacts that new users
have with autotest, hence, we need to keep them well maintained
and up to date. This patch gives an overhaul on the existing
samples, updates them with new APIs and make them follow the
autotest control file specification.

Some control files showing up simple features were merged
with more complete ones, so we hopefully end with something
more organized and understandable.

The plan is to do the same with the server side ones, fixing
eventual problems found. Will also try to keep testing them
against the latest trunk/git HEAD so we try to keep
regressions at a minimum.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4905 592f7852-d20e-0410-864c-8624ca9c26a4