pass SSH options into SSHHost
We found the previous patch only fixed a part of our problem. We just found
there are occurrences inside server test control files which called
hosts.create_host(hostname) without properly passed in ssh options. And I do
not want to touch the test control files at all.

So here is a solution, my previous changes to SSHHost is reverted in this
patch, since it is not necessary any more.
I tested this change with both client and server side sleeptest against a
remote VM ( on port other than 22).

Signed-off-by: Eric Li <ericli@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4944 592f7852-d20e-0410-864c-8624ca9c26a4