Replacing regex search operations with precompiled regex objects
Currently there's a mix of regex usage styles in autotest. Precompiled
regex should be preferred, specially when they are used many times.

Note 1: "many times" means not only multiple ocurrences in a source
file, but specially many times during the application execution, so
code in common_lib is very relevant.

Note 2: Python maintains a limited cache for not previously compiled
regex, but we should not count on cache hits, specially as autotest grows.

Signed-off-by: Cleber Rosa <crosa@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4953 592f7852-d20e-0410-864c-8624ca9c26a4