error: Introduce exception context strings
In complex tests (KVM) an exception string is often not informative enough and
the traceback and source code have to be examined in order to figure out what
caused the exception.  Context strings are a way for tests to provide
information about what they're doing, so that when an exception is raised, this
information will be embedded in the exception string.  The result is a concise
yet highly informative exception string, which should make it very easy to
figure out where/when the exception was raised.

A typical example for a test where this may be useful is KVM's reboot test.
Some exceptions can be raised either before or after the VM is rebooted (e.g.
logging into the guest can fail) and whether they are raised before or after
is critical to the understanding of the failure.  Normally the traceback would
have to be examined, but the proposed method makes it easy to know where the
exception is raised without doing so.  To achieve this, the reboot test should
place calls to error.context() as follows:

error.context("before reboot")
<carry out pre-reboot actions>
error.context("sending reboot command")
<send the reboot command>
error.context("after reboot")
<carry out post-reboot actions>

If login fails in the pre-reboot section, the resulting exception string can
can have something like "context: before reboot" embedded in it.  (The actual
embedding is done in the next patch in the series.)

Differences from previous version:
- Allow for 2 context levels per function using base_context().  This may be
  useful for some functions.
- Update the comment in error.py.
- Rename context_for_exception() to exception_context().
- Add set_exception_context().
- Add join_contexts().

Signed-off-by: Michael Goldish <mgoldish@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@5120 592f7852-d20e-0410-864c-8624ca9c26a4