remember used harness after reboot
If autotest client is run with custom harness, its name is stored
in state file. This allows autotest to continue (e.g. after reboot)
with harness that was specified on first run.

If "--harness" option is present on command line, this
option is used regardless of what is stored in state file.
Added log info about used harness.

Changes from v1: Put the harness debug statement inside the harness
module
Changes from v2: Fix unittest breakage this test was causing. The
underlying unittest code was expecting the default value for the
harness to be None rather than an empty string.

Signed-off-by: Jan Stancek <jstancek@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@5209 592f7852-d20e-0410-864c-8624ca9c26a4