Recognize parameters surrounded by quotes
Sometimes we need to pass the parameters which contains blanks and it
is usually done through surrounding it with quotes. So this patch use
re.findall() to try to recognize the parameter surrounded by quotes
and pass it as one single argument to the control file.

Kvm subtest would first benefit from this patch as we could pass
something like "vms='vm1 vm2'" to it.

Changes from v1:
- Isolate the parsing code to a specific classmethod.
- Create a simple unittest for the parsing code.

Signed-off-by: Jason Wang <jasowang@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@5213 592f7852-d20e-0410-864c-8624ca9c26a4