Fix check_diskspace to work at a granularity less than 1GB
This was breaking a check we have that there be 100MB free
on / by making it always appear true as it could no longer
measure a value that low.

The previous change inadvertently limited the granularity of the disk
space to check to whole integer GB rather than fractional.  This gives
it a granularity of MB so that a test for gb=0.1 works properly again.

Signed-off-by: Gregory Smith <gps@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@5236 592f7852-d20e-0410-864c-8624ca9c26a4