[PATCH] After_iteration hooks are not executed if test fails.
Base_test's _call_run_once() does not execute the after_iteration hooks
if the test fails.  It is important to execute the after_iteration hooks
in the event of a test failure since these hooks may be used to gather
important test logs to help diagnose the failure.  This patch
modifies _call_run_once() to execute after_iteration hooks regardless of
the outcome of the test.

A new unit test was added to verify the behavior of _call_run_once().  The
patch was tested using unittest_suite.py to ensure that it did not break
other functionalities.  It was also tested with an actual after_iteration
hook that collected test logs and the test was forced to fail.

Risk: High
Visibility: The order of when the after_iteration hooks has been changed
and may affect existing tests.
Bugs: None

Signed-off-by: Thieu Le <thieule@chromium.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@5271 592f7852-d20e-0410-864c-8624ca9c26a4