web front end: Change the atomic group pull down menu
Change the atomic group pull down menu displayed in the admin interface
for Labels to a text box in hopes that it won't prepopulate the
selection with a value that then inadvertently gets set when the field
is intentionally NULL (the common case).
Also: display the atomic group in the labels list view.

Signed-off-by: Gregory Smith <gps@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@5278 592f7852-d20e-0410-864c-8624ca9c26a4