Adding 3G wireless modem firmware check in HWID test
Change-Id: If1ffbb9597642602f914716d934a241fdbe99709

BUG=chrome-os-partner:2952
TEST=It should pass HWID validation test only when version_3g_firmware matches

This change is to detect the firmware version of Samsung Y3300 which supports USB suspend. It can check Qualcomm firmware version, but it may be used optionally. I verified it on DVT2 with the current Y3300(before suspend fix).

When we get the final firmware version from Samsung, we need to update component file to check the firmware ID.

Review URL: http://codereview.chromium.org/6770001