network_3GSafetyDance: Ensure Favorite=True before setting AutoConnect=False (R14)
If we run 3GSafetyDance on a newly-imaged machine, we've never
connected to Verizon, so Verizon is not set as a favorite service, so
flimflam won't let us set AutoConnect on it, so when we try to disable
AutoConnect, flimflam gives us an InvalidService error.

This patch adds an extra workaround that makes sure Favorite=True
before trying to set AutoConnect=False.  (It also asserts that the
final state of the service is what we expect.)

(This is a cherry-pick of http://gerrit.chromium.org/gerrit/#change,5679
and http://gerrit.chromium.org/gerrit/#change,5683 in R15.)

BUG=chromium-os:18981
TEST=Adhoc; reimaged a machine and ran SafetyDance, and it passed

Change-Id: Ia713e4aae7225d5cb0a2a966d239da6f7bd3d164
Reviewed-on: http://gerrit.chromium.org/gerrit/5749
Tested-by: ttuttle <ttuttle@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>