Don't directly invoke smoke suite, use packaging.
Currently when the AU test runs, it picks up the smoke suite control
file from the Autotest server instead of packaging. Due to this we end
up with a version mismatch on occasion.

I've switched the AU test to pull the smoke suite via packaging and
execute the extracted control file. Doing so ensures we run the right
version with each test.

Will port this and the previous parser URL fix back to R15.

BUG=none
TEST=Ran control file. Tested on production server.
http://cautotest/afe/#tab_id=view_job&object_id=67455

(cherry-picked from R16, http://gerrit.chromium.org/gerrit/#change,8751)

Change-Id: I72b77aefe3b893a9a7507cb6f00ce653094ca9a5
Reviewed-on: http://gerrit.chromium.org/gerrit/9862
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Tested-by: Dale Curtis <dalecurtis@chromium.org>