(factory-980B) factory: Add "hwid_cfg" to mark test image.
For RMA images.
Reviewed-on: https://gerrit.chromium.org/gerrit/11122

The same factory test image is used for factory/RMA/... (may be more in future),
so we need some mechanism to differentiate the type of test we performed.

The HWID already has configuration names to describe what we are testing for, so
this CL first tries to look up "cfg" from installed HWID folder, and then decide
which test list to use; and in the end add that to "report_tag".

BUG=chromium-os:22460
TEST=manual

Change-Id: Idb09e083920044562e1a256d0a128e50fd4ae869
Reviewed-on: https://gerrit.chromium.org/gerrit/11257
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>