factory_ExternalStorage: identify device using only subsystem.
*** This CL is only for factory branch ***
Current way to detect SD card fails with the SD card reader we have. We
can identify device using only subsystem='block'. Since we test SD card
lock switch, one cannot pass this test with a USB stick.
This change serves as a workaround for now. Need to discuss if we
actually want to change SD card test this way.

Signed-off-by: Vic Yang <victoryang@chromium.org>

BUG=chrome-os-partner:8597
TEST=Tested on Alex.

Change-Id: I4a9e1440318c8053367c339c8f1f6a903a7fdd60
Reviewed-on: https://gerrit.chromium.org/gerrit/18768
Reviewed-by: Vic Yang <victoryang@chromium.org>
Tested-by: Vic Yang <victoryang@chromium.org>