factory: Fix camera detection bug and export some params for cam test
The CL fix a rare condition that the program can't detect the camera
after repetitive AB panel test runs because the handler of the camera in
the previous run is not properly garbage collected. It also exports more
test parameters in the camera performance test library so that the
developer can adjust their values more easily down in the factory floor.
Finally, the performance test library now pre-normalize the point
coordinates more appropriately before doing the stratified random
sampling.

BUG=chrome-os-partner:6617
TEST=Test it in controlled environment(fixture) -> works

Change-Id: If3a24928c92840e7c903abe58bb7ce2bb7bbc76c
Reviewed-on: https://gerrit.chromium.org/gerrit/19056
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>
Reviewed-by: Tai-Hsu Lin <sheckylin@chromium.org>