factory_AudioLoop test: Allow test combination be blacklisted
Adds a list of blacklisted test combination of (channel, input dev)
to the test argument, and skips them in the main test.
The purpose to blacklist some combinations is to workaround factory
fixture and environment's limitation.
For example, left speaker is placed far away from dmic with a big
fan in between, in that case we are okay to skip certain combination
and test the rest to cover all component's functionality.

BUG=chrome-os-partner:16419
TEST=Manual run audioloop test with some teste combination black listed.

Change-Id: Ifb911d6f3a34f94b4107322bd731712649de97f6
Reviewed-on: https://gerrit.chromium.org/gerrit/38795
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>