kernel_ConfigVerify: fix report of "unset" value
When the desired state was a config to be disabled, reporting was trying
to make a string out of None. Instead, do this manually to match how the
config looks in a kernel's .config file ("unset").

BUG=chromium:342951,chromium:341583
TEST=daisy_spring fails cleanly when using DEBUG_SET_MODULE_RONX

Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/185966
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>

Change-Id: I697cc5acdd68feb1c0e3cb08f0c4ae337ad46ad2
(cherry picked from commit 6e1ac69eeca478c14be29b4710dc1cce27d45e67)
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188371
Reviewed-by: Will Drewry <wad@chromium.org>