Add firmware_EventLog test to verify event log generation and timestamps
This patch adds a new faft_lv4 test to validate the functionality of the
firmware event log. It will boot the system in normal, developer and
recovery mode, verifying that the right events are logged for each. It
also tests a suspend/resume cycle, allowing either the ACPI events used
on x86 boards or the userspace Sleep/Wake events introduced with Spring
and Pit. (We can't check for 'FW Wake' because not all not-x86 boards
have a resume component in firmware.) The events are correlated to the
test section they were generated in by their timestamps, so this also
confirms that accurate timestamps are generated.

CQ-DEPEND=CL:238467
BUG=chrome-os-partner:38607
TEST=Tested on Falco (ToT), Blaze (ToT) and Pi (release). Pi failed in
the expected manner because U-Boot is missing RTC support (and all
timestamps are added as 2000-00-00). Blaze failed because its
timestamps' months are off by one... turns out the RTC driver code was
incorrectly copied from the 0-based-month-using kernel to our
1-based-month-using coreboot.

Change-Id: I2760080030468ed71dd0b7359ac69a269f6ab47a
Originally-signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238078
(cherry picked from commit 9ece2c6fb01d4c29d23a7238def15bdf656c885c)
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/264166