Add hold_cold_reset back into DEFAULTS.py
hold_cold_reset was removed in ToT by CL:270765, but not in the
firmware branch. Unfortunately, it appears to have been accidentally
clobbered in the firmware branch when CL:295260 was cherry-picked.

CL:270765 has dependencies, so this patch simply adds hold_cold_reset
back in so that we don't need to cherry-pick additional patches.

BUG=none
BRANCH=firmware-veyron only
TEST=firmware_UserRequestRecovery works in firmware-veyron branch

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I229757099fcb7a7ee08aacd15982390e5a53a940
Reviewed-on: https://chromium-review.googlesource.com/296492
Reviewed-by: Julius Werner <jwerner@chromium.org>