Stress Runs: Add an additional timeout for SESSION_STOP to login_LoginSuccess.
This test has been failing during stress runs with 50 iterations. It is
failing due to a timeout. We are extending this timeout so that we can allow the test to
fail for other reasons as it is an important test for FSI readiness.

TEST=test_that -b <board> <ip> login_LoginSuccess
BUG=540015

Change-Id: I60ba745eaa7a0ec8bb3b14454bb7a526c7708e48
Previous-Reviewed-on: https://chromium-review.googlesource.com/303952
(cherry picked from commit fd4584c7c2879a6252a3c855111f0e080abbfa83)
Reviewed-on: https://chromium-review.googlesource.com/305338
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>