graphics_WebGLPerformance: rev sources.
This revs the tests to commit 3729e8afac99:
http://hg.mozilla.org/users/bjacob_mozilla.com/webgl-perf-tests/raw-file/3729e8afac99/index.html

In addition fix average computation/log(0) handling. 0ms should
basically be treated as 1ms for geometric average computation.
Now practically with the new revision 0ms doesn't even happen on
Samus, so we are kind of good for a while. If this happens with a
faster machine again we can always multiply |repeat| with 10 or so.

BUG=chromium:548352
TEST=Ran on lumpy, link, samus.

Change-Id: Ie75b314979dc75e379b3d8788a91ad6f1e5d7059
Reviewed-on: https://chromium-review.googlesource.com/309517
Commit-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309912