[Cros WebRtc] Publish specific errors in the test error message.
Although the specific cause of test failure was known, we were
logging that in a file but the test error message that is shown
in wmatrix remained generic.

To get an immediate impression of what caused the test failure,
I'm reporting the error message as the error message.

BUG=None.
TEST=Passes locally.

Change-Id: Ic7e49229b3f563ccbeddd980071c4c7a6f1dfb35
Previous-Reviewed-on: https://chromium-review.googlesource.com/327321
(cherry picked from commit 52b8f748a64091aa4d8747308a3fed3706a1bd67)
Reviewed-on: https://chromium-review.googlesource.com/327215
Reviewed-by: Christian Paulin <cpaulin@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>