firmware_ECBootTime More robust detection
With the increase in speed of the EC there is more chance of the
messages being dropped due to the finite size of the TX buffer.
The ECBootTime test is very sensitive to messages being dropped.
This patch makes things a little more robust. It looks for the
two UART messages so we know we are on the second boot, and then
gets the first timestamp it sees rather than lookink for a timestamp
on a message it knows only occurs in the second boot and not the first
one.

Of course the correct thing to do is scrap this test (which doesn't
really measure anything) and write a replacement.

TEST=pass firmware_ECBootTime
BUG=chrome-os-partner:48564

Change-Id: I9c113378163dec8a23ea94f9433e49d2dbb3632a
Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/332525
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Tested-by: Divya Jyothi <divya.jyothi@intel.com>
Commit-Queue: Divya Jyothi <divya.jyothi@intel.com>