chrome_binary_test: add link for component build.
On 32 bit systems we use chrome with component build now. That means
the *unittest binaries are smaller, but can't find their dependency.
Use the setup in chrome_binary_test to create a symlink to
/opt/google/chrome/lib if needed. Notice this requires a binary match
between the installed chrome and unittests. Emit a reminder to the
user that mixing and matching isn't meaningful anymore.

Also minor cleanup.

TEST=test_that $DUT security_SandboxLinuxUnittests ideo_VideoDecodeAccelerator.h264 video_VideoEncodeAccelerator.h264
BUG=chromium:606373

Change-Id: I297a2b93bdb7e3443b40f459f15270511f505da0
Reviewed-on: https://chromium-review.googlesource.com/340894
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>