[autotest] Set container's hostname to the dut hostname
Update the hostname of the test container to be `dut_name`. Some TradeFed tests
use hostname in test results, which is used to group test results in dashboard.
The default container name is set to be the name of the folder, which is unique
(as it is composed of job id and timestamp. For better result view, the
container's hostname is set to be the dut hostname.

BUG=chromium:637467
TEST=local verify

Change-Id: I1481d88a890020af634a499fb6d469574f9e4504
Reviewed-on: https://chromium-review.googlesource.com/376221
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/406737
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Mohan Konduri <krk@chromium.org>