[autotest] Keep servo repair events out of status.log in tests.
For tests that depend on servo, and use the standard recommended
boilerplate for ensuring that the servo host is created, the servo
host's repair() method will be called prior to the test.  The
resulting log records in status.log cause wmatrix to report bogus
tests that pollute the UI.

This changes servo repair in this case not to write to status.log.

This change is a cherry-pick from the main branch, and includes a
required bug fix from a separate CL.

BUG=chromium:643703,chromium:661201
TEST=unit tests

Change-Id: I64756f20c1eda0d60d16a06bb1838991c22f4f3c
Reviewed-on: https://chromium-review.googlesource.com/411948
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>