Autotest: change gs_offloader to seperate upload age from delete age.
Currently gs_offloader has a single param to select how old a job
should be before it is uploaded and deleted.

Introduce two new params one to select how old the job is when it is
uploaded, and a second when it is deleted.

If these params are the same or the original param is set the
functionality of gs_offloader should be the same.

BUG=chromium:659202
TEST=local testing on moblab, ran/fixed available unit tests.

Change-Id: I91e63ba39ddbd8c0ea408c3f5111eead76eaa0f7
Reviewed-on: https://chromium-review.googlesource.com/410060
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>
(cherry picked from commit 5ba5fb86d92a242691def5fc87ff35e51497b2ae)
Reviewed-on: https://chromium-review.googlesource.com/412022
Commit-Queue: Keith Haddow <haddowk@chromium.org>