cheets_GTS: retry if failed to connect to DUT.
On a number of runs of the new GTS it is failing to find a connected adb
device in time and is failing before the test runs. This adds a single
retry and improves the error messaging that is currently being shown on
a failure.

I have not been able to reproduce it locally but have verified this by
hacking the results object to have a failed test's output.

BUG=b:34671162
TEST=test_that <ip> cheets_GTS.4.1_r1.GtsAdminTestCases

Change-Id: I568804849f949357c3ec814d1421b1a22da71adc
Reviewed-on: https://chromium-review.googlesource.com/431806
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434956
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>