[autotest] Add setting to allow devices to be in dev mode.
A new check was added to the cros device repair to ensure all
devices must not be in dev mode.  This does not apply to autotest
so add a config setting that moblab can use to bypass this repair
check.

TEST=dev build flashed to moblab and cts test run.
BUG=chromium:712374

Change-Id: I455c3ca8b9a0a1b3bb1a04f3b94983d206e0e6ce
Reviewed-on: https://chromium-review.googlesource.com/480912
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/487222
Commit-Queue: Keith Haddow <haddowk@chromium.org>