[autotest] Change the public cts/gts to allow test to be specified.
Users when creating a new suite can optionally pass in a list of
tests.  Rather than schedule all the tests in the suite the suite
will consist of just the tests in the suite that the user specified.

This allows quicker re-runs of failing tests by partners on moblab.

TEST=tested locally on moblab.
BUG=chromium:725601

Change-Id: I076084847a28ddc1bea52cb43e49a3edf8efad16
Reviewed-on: https://chromium-review.googlesource.com/513602
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Michael Tang <ntang@chromium.org>
Commit-Queue: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>