[moblab] Add new featutre to run suite to limit the tests run.
When partners re-run a CTS suite they often just want to run a
subset of the jobs.  The UI now lets you specify a list of test
names.  The suite will only run the tests that are in the suite
and in the specified list.

Demo on http://100.107.3.16/moblab_setup

TEST=local moblab
BUG=chromium:725601

Change-Id: I7708d2f4765370bff130e224dd0d33854202f6fd
Reviewed-on: https://chromium-review.googlesource.com/514365
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Jinsong Mu <jinsong@google.com>
Commit-Queue: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>