cheets_CTS: Retry on "Error: No end message for the last chunk."
This happens when tradefed failed badly in the middle of the test and
did not print the pass/fail/notexecuted count at the end of the aborted
test run.

This CL regards such a chunk of tests to be not-executed and report as such.
The subclass, cheets_CTS_N.py, will retry if it sees the new parse result.

BUG=b:64095702
TEST=./utils/unittest_suite.py server.cros.tradefed_test_unittest

Change-Id: I3ed6fcab21174b255dd640efb82cb366161bf296
Reviewed-on: https://chromium-review.googlesource.com/593378
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
(cherry picked from commit 24a12c39062e30a70cef1a004b4f71571d46036f)
Reviewed-on: https://chromium-review.googlesource.com/602167
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>