tradefed_test: Simplify log collection by using symlinks.
Previously we copied the results and logs from the tradefed's directory
to where our tools can recognize. This CL simplifies the steps by introducing
symlinks from the former to the latter.

Advantages are:
 * Less code, hence less bugs. We no more need to parse stdout for obtaining
   the directory name, nor to run explicit tree copying.
 * (Partial) log collection even in the case of aborted tradefed runs.
   By using symlinks instead of copy-after-finish, as long as tradefed flushes
   out, the logs upto the point of timeout abort etc can be collected.
   The logs are, even if they are partial, useful for debugging.

Drawbacks are we rely on an internal implementation of the tradefed tool that
it can transparently deals with symlinks injected at this location. That said,
it is working well for current cheets_CTS_N, cheets_CTS_P, and cheets_GTS, and
since we are anyway relying on the directory structures, etc, the new assumption
won't be a big deal.

BUG=b:73136256
TEST=run some cheets_CTS_N tests with/without manually shortening the timelimit

Change-Id: I6c101e94ee9ede15f6cd38e29ecb2d974de20c15
Reviewed-on: https://chromium-review.googlesource.com/937144
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
(cherry picked from commit 4c7c3000238ffa2e942c38a9f5076412e631a2d1)
Reviewed-on: https://chromium-review.googlesource.com/1027198
Tested-by: Ilja H. Friedel <ihf@chromium.org>