hostap_config: Replace mutable list argument with None
This fixes the following pylint warning:
Dangerous default value {} as argument (dangerous-default-value)

From the python docs:
(https://docs.python.org/3/tutorial/controlflow.html#default-argument-values)
The default value is evaluated only once. This makes a difference when the
default is a mutable object such as a list, dictionary, or instances of
most classes. A function with a mutable argument accumulates the arguments
passed to it on subsequent calls.

BUG=chromium:847272
TEST=network_WiFi_Prefer5Ghz as a sanity check.

Change-Id: I4f9c2f1dcdb1603f8ba1e13f6e4eaf85d71227fb
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1640319
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>