autotest: servo keyboard commands now run through set/get
In the dependent CL the rpc calls for specific keyboard commands are
changed to now use the standard servod set/get methodology.
This means that these commands need to be refactored to now use that
system as well.

Also remove d_key as that is no longer supported in servod after
crrev.com/c/1550660

CQ-DEPEND=CL:1158322

BUG=chromium:869335
TEST=None

Change-Id: I04a7ba6030569e561279aa5e1766f6fd75829640
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1158444
Reviewed-by: Garry Wang <xianuowang@chromium.org>