[autotest] tcpdump_analyzer: don't require rate info for packet capture
If we're running a packet capture on the same radio as the AP (e.g.,
without a separate packet capture, and only a single radio that supports
the frequency band in use), then the packet capture might not contain
rate info in the radiotap header. We don't actually use the rate info in
most cases, and if we really cared about integrity, we'd just use the
FCS checks instead.

So, drop this check, and capture the frames regardless. I leave the
debug message there, in case we have trouble with these captures in the
future.

BUG=chromium:774808
TEST=network_WiFi_ChannelScanDwellTime on Gale

Change-Id: Icf0cb815f6d32d32f237453637fade75b12ca6de
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1159466
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
(cherry picked from commit c02f98dba0f5ace9edab1e1664655c3b21cbd84d)
Reviewed-on: https://chromium-review.googlesource.com/1174673