Bluetooth: Modify check for pairable property
This test assumes that Bluetooth adapter Pairable property
will be true after it a power cycle, which is not a valid assumption.
This change modify the test so that it doesn't hold that assumption.
Changes from CL 1435487 has been cherry picked into this CL.

BUG=chromium:925135
TEST=Tested locally

Change-Id: If51198ed820b871a147f801f57ac7a7925d3605b
Reviewed-on: https://chromium-review.googlesource.com/1440190
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>