faft: Allow falsifying RO and RW fwids, for BIOS and EC.
In order to test the results of firmware updates, we need to verify
that certain sections of firmware have actually been modified.
The simplest way to do that is to alter the FWIDs.

The get_fwid and get_ecid functions now return values consistent with
the new modify_fwid and modify_ecid functions.

The get_fwid function definition is now in the flashrom handler class,
for code reuse.

BUG=chromium:967917
TEST=Run firmware_FAFTSetup

Change-Id: I96910e63df06e6b7353bff3385713f409680eb95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1626619
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>