network_WiFi_Reset: perform WiFiCellTestBase cleanup before reboot
If we fail the test, we might reboot the DUT. Run the parent cleanup
before forcing a reboot, since the parent is trying to maintain an
xmlrpc connection for doing some of the cleanup.

This is not the true cause of $BUG, but it does confuse the log results
a bit.

BUG=chromium:900449
TEST=network_WiFi_Reset, and force a DUT failure

Change-Id: I639a07c63d4e8dfd511e03fb4cfffdbeaf205c8d
Reviewed-on: https://chromium-review.googlesource.com/1572481
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Neeraj Poojary <npoojary@google.com>