pseudomodem: simplify subscription state code
The unregistered state was always UNKNOWN, so in the interest of
simplifying the code before refactoring it, we can remove the
unregistered subscription state from the test APIs. This will
be replaced with an empty PCO list when we start using PCOs.

BUG=chromium:928572
TEST=cellular_OutOfCreditsSubscriptionState,
  cellular_ActivateLTE without ActivationCompleteTest

Change-Id: I0220ffc0321270b801ff315238584be24867b14c
Reviewed-on: https://chromium-review.googlesource.com/1536629
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>