firmware_Cr50ConsoleCommands: Support Wilco
The Cr50 in Wilco has some unique GPIO configuration that is reflected
in the outputs of the gpiocfg and pinmux commands. Account for those
differences in firmware_Cr50ConsoleCommands and remove an unnecessary EC
access that would cause problems on Wilco. Also, make some of the output
more readable.

BUG=b:129421548
TEST=Ran firmware_Cr50ConsoleCommands on Arcada, Liara, and Bob

Change-Id: If240ea841d482d67f1726f8483ea90f2d211ffef
Reviewed-on: https://chromium-review.googlesource.com/1546172
Commit-Ready: Abe Levkoy <alevkoy@chromium.org>
Tested-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>