cellular_ActivateLTE: remove ActivationComplete sub-test
This test no longer seems to describe how the network operates,
since networks will allow you to register before activation but
send a PCO that you are not subscribed. The other sub-tests in
this test already account for this behavior so we can get rid of
this artifact.

BUG=chromium:928572
TEST=run test

Change-Id: If627a0578e00cd5df964a64dfad6da284d70def5
Reviewed-on: https://chromium-review.googlesource.com/1551968
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>