Clean up telemetry_Crosperf and handle failed tests
The `run_benchmark` command no longer needs to be run in background
since profiling has been integrated into it. Cleaned up all codes
related to background command.

Meanwhile, when the command runs correctly but the test fails, the
command will return 1 other than throwing an error. So we also want to
handle it so that the error will not be hidden.

TEST=Tested with several telemetry_Crosperf tests.
BUG=chromium:950064

Change-Id: I71635bcf7f72240d9a82585a7617632b67629c35
Reviewed-on: https://chromium-review.googlesource.com/1554995
Commit-Ready: Zhizhou Yang <zhizhouy@google.com>
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: Caroline Tice <cmtice@chromium.org>