skylab_suite: stop pulling dut name from task_spec
This change includes some behavioral changes to the way in which success
or failure of provision suite is logged and displayed on the waterfall.
In particular, success or failure links will no longer include the dut
name. Follow-up TODOs are left behind to restore that functionality, in
a way that doesn't depend on the tast request being dut-specific.

BUG=chromium:946787
TEST=None

Change-Id: I1f11059e8dbaa7ff05daad31abe5577ebd5ac7e4
Reviewed-on: https://chromium-review.googlesource.com/1555541
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>