gbmtest: Allow test to work even when no display is connected
Historically we have tried to find the first drm device with a connected
display, and if we can't find any we don't run the test at all. However in the
case of chromeboxes in the lab this means the test doesn't run. This change
allows the test to run these as well.

BUG=none
TEST=this is a test, so I ran it

Change-Id: I310a725db5360de2f25f961f2e14649a4ca3192d
Signed-off-by: Stphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1555733
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>