network_WiFi_BT_AntennaCoex: Use ch11, not ch13
2472 MHz aka channel 13 is forbidden in the US. This test
tries to connect there in the very last sub-test. That will fail
because we never even scan on that channel. Change it to channel 11.
Note that the failure only happens on DUTs which have a US regdom,
as the world regdom (country code:00) will allow scanning on ch13.

You can notice this as the test failure is:
Expected connection to BT_AntennaCoex_d_gw1o2_ch13 to succeed, but it
failed with reason: FAIL(Discovery timed out).

And the wpa_debug shows the scan list never included 2472:

2019-04-05T09:38:53.707821+00:00 DEBUG wpa_supplicant[733]: nl80211: Scan \
included frequencies: 2412 2417 2422 2427 2432 2437 2442 2447 2452 2457 \
2462 5180 5200 5220 5240 5260 5280 5300 5320 5500 5520 5540 5560 5580 5600\
5620 5640 5660 5680 5700 5720 5745 5765 5785 5805 5825

BUG=chromium:876920
TEST=Lock chromeos15-row2-rack2-host1 that was failing this test
in the lab, run test with this change and watch it pass.

Change-Id: I1966d25146952f7278a81d79fd9c6db28cd8bf0a
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1562531
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>