sarien: Avoid powering off AP during CCD open
On Wilco, the power button presses used to confirm physical presence
during CCD open are long enough to power off the AP, which interferes
with tests that exercise this behavior. To work around this, stop powerd
on the AP before running gsctool and press the power button to turn the
AP back on, if it is turned off while running 'ccd open' from the Cr50
console.

BUG=b:127694512
TEST=Ran Cr50Password and Cr50Unlock on Arcada

Change-Id: I5d9e171e7761254d29a9e9b99b4bfb4ee033466c
Reviewed-on: https://chromium-review.googlesource.com/1519072
Commit-Ready: Abe Levkoy <alevkoy@chromium.org>
Tested-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>