arc: disable cgroup restrictions for tests by default
We're planning on making the cgroup controls over ARC much more
restrictive for the amount of CPU that can be used when ARC is in the
background, but this makes autotmated testing much slower so we've
added a flag to Chrome to disable this.  This adds the flag for tests
that run with ARC enabled, but adds an option so that a test that
wants to run with the restrictions can do so.

BUG=b:118696077, b:124756345
TEST=run cheets_AuthPerf

Change-Id: If8e3c3eb1c7dd28e546555f7a3050ca9f14a5128
Reviewed-on: https://chromium-review.googlesource.com/1566163
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>