[Autotest] policy_DefaultSearchProviderEnabled
Testing the DefaultSearchProviderEnabled policy.

If the policy is set to True/None then typing search queries in the
omnibox will result in searching on google.com.
If the policy is set to False then typing search queries in the
omnibox will not result in searching on google.com.

BUG=None
TEST=this

Change-Id: I724d35393a019deb1edd1c2490b09ec26fbe81c6
Reviewed-on: https://chromium-review.googlesource.com/1566442
Commit-Ready: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>