firmware_DevScreenTimeout: Add a loop to check the availability of /tmp/firmware-boot-time, with a timeout
The test item is need to read /tmp/firmware-boot-time immediately after the system is available on the network,
but this file is not created until the send-boot-metrics upstart job.
It always to check the availability of /tmp/firmware-boot-time until timeout.

BUG=b:128552848
Branch=None
TEST=Ran firmware_DevScreenTimeout passed on Nautilus.

Change-Id: I9cd868361de77be5f817ca13e7400dd1c63a2de4
Signed-off-by: raymondchung <raymondchung@ami.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/1567581
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Raymond Chung <raymondchung@ami.corp-partner.google.com>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>