[Autotest] policy_DefaultSearchProviderKeyword and policy_DefaultSearchProviderSearchURL
Testing the DefaultSearchProviderKeyword and DefaultSearchProviderSearchURL policies.

When the
DefaultSearchProviderSearchURL policy is set, the specified search url
will be used when a value is entered in the omnibox. When the
DefaultSearchProviderKeyword is set, the value will trigger the shortcut
used in the omnibox to trigger the search for this provider.

BUG=None
TEST=this

Change-Id: I256a4a7083f8dece809e6368cb89875e3f56ba8f
Reviewed-on: https://chromium-review.googlesource.com/1569685
Commit-Ready: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>