hostap_config: re-enforce HT requirements
Pure N should require HT support. We don't want to continue handicapping
ourselves for daisy/snow. (Or at least, we shouldn't keep whining about
this condition.) It happens that all supported devices now work fine
with require_ht, so let's turn it back on.

Per hostapd, "require_ht" and "require_vht" mean including the HT or VHT
BSS selector in the Supported rates (or Extended supported rates) IE,
which means that the client station must support HT or VHT in order to
associate. It does *not* mean that the AP won't also advertise legacy
(non-(V)HT) rates, and so clients are still allowed to regress from HT
to legacy rates.

Given the above, it's not 100% clear how we should translate 11xx_PURE
into hostapd properties. Particularly, 11AC_PURE could be interpreted to
mean:

* require VHT or
* require HT and VHT

I've selected the former for now, as it's the least churn, but we can
update this choice later if desired.

BUG=chromium:237370
TEST=run suite:wifi_matfunc on daisy_skate and others

Change-Id: Iade412b5b0cdac3950f9425d4145874a44e3413f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1579123
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>