[Autotest] Adding new autotest for policy_SecondaryGoogleAccountSigninAllowed.
Testing the SecondaryGoogleAccountSigninAllowed policy.

If the policy is set to True/Not Set then users can sign in to Chrome
with multiple accounts. If the policy is set to False then users won't
be given an option to sign in with more than one account.

BUG=None
TEST=this

Change-Id: Id42e5e9148ae5af66e902a7d78e972306dbc6727
Reviewed-on: https://chromium-review.googlesource.com/1576140
Commit-Ready: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>