network_WiFi_APSupportedRates: test against APs without legacy rates
Many APs today (esp. in enterprise configurations) might disable legacy
rates. We want to play nicely with them. Test that, when an AP
intentionally removes lower rates from its {Basic,Supported}Rates IEs,
that the DUT respects that.

Note that this does not yet pass on many devices. It's useful to have
this test running in the lab, so start it in suite:wifi_flaky, and we
can decide later when this is ready for wifi_matfunc.

BUG=chromium:953702
TEST=run test, with and without preliminary driver fixes

Change-Id: Iabde2278837ea14704df84904fa6517076b237f9
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1579125
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>