servo_LabstationVerification: add integration test for labstations
This test implements an end-to-end verification of servod on a
labstation image to validate images before pushing them out to the lab.
It covers
- basic servod controls
- testing the consoles
- imaging process in the lab

Most importantly: it tests the labstation itself, and is DUT
independent. Thus is can be used in reverse as well: if the labstation
image is qualified, it can be used to determine whether a new DUT will
face issues during deployment or once in the lab.

Also important: the test provides a basic mechanism to run a variety of
targeted tests against the servo<>DUT combination. So it provides the
infrastructure to enhance and improve labstation validation testing as
we move forward, as new tests just need to hook into this test.

BUG=chromium:939462
TEST=manual testing of the sub steps, but requires merge to test E2E
given the devserver issues.

Change-Id: If6c03c8e12b7cefa2d0286ab4b2b95f9fc60d9ac
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1579743
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>