servo_LabControlVerification: verify basic lab controls on labstation
This test runs on a device acting as a servo host to validate that
among other things hdctools can query critical lab controls, as well as
query all consoles.

The setup for this requires a chrome os device that acts as the servo
host (that's the device the test is run against), a servo device
attached to that chrome os device, and a reasonably well behaved chrome
os devices attached to the servo that only acts to respond to servod
controls.

BUG=chromium:939462
TEST=manual testing
test_that --autotest_dir [...] $IP servo_LabControlVerification
...
[Passed]

(setup here is a wukong, ccd v4, and a nocturne)

Change-Id: I9885265941271b02284a59b26ea67021318ae68a
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1585981
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>