servo_v4: expand charge utils
This change expands the servo v4 charging utility to run a more thorough
check when changing states. It
- verifies that a change PD src/snk to the other is properly recognized
on v4
- there is power provided to the DUT (through ADCs on the v4)
- the EC on the DUT side recognizes the right state

It also adds a recovery mechanism that essentially on failure to set the
v4 to a certain role, reverts to the previous role, before reattempting
to set a given role. This has shown to be an effective way to recover
the v4's PD to the DUT.

BUG=b:120233091
TEST=testing with power_ServoChargeStress

Change-Id: Ic4859cb1963d640591ed42989668c086e425e85f
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1586624
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>