power_Monitoring: add a power_ServodWrapper loop-runner
This change introduces power_Monitoring, a test that runs tests in a
specified suite wrapped with power_ServodWrapper for a given amount of
time. The test ensures the DUTs do not run out of power by using v4's
power management to keep the DUTs always above a specified threshold. It
charges the DUT in those cases to an upper threshold before continuining
the tests.

As a demonstration vehicle, this CL adds power_LoadTest.1hour to the
suite. A subsequent CL adds the first batch of tests to this runner

BUG=chromium:923765
TEST=manual testing
test_that --autotest_dir=[...] $DUT_IP power_Monitoring

observe as tests start looping

Change-Id: Ib147f127fbb70730716960d2a997fd6a4a73fb36
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1423938
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>