[autotest] Fix ArcVideoCaptureAllowed
Some fixes to reduce the failing of ArcVideoCaptureAllowed.

This CL currently adds a long sleep. There are much more complicated
ways to get around this, but for now this will test if this fix will
work in the lab. It has been tested on both local/lab machines and looks
OK.

TEST=this
BUG=None

Change-Id: I9026551e437aa00efb6eb859235d8dd864d0dd83
Reviewed-on: https://chromium-review.googlesource.com/1587482
Commit-Ready: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>