servo: explicitly enable the keyboard
Due to recent changes on servod, the keyboard handlers have to be
initialized explicitly. This happens automatically for onboard
keyboards, but by default the usb keyboard is off.
However, for some devices (boxes, etc) FAFT has to run using the usb
keyboard through servod. While there is support in servod for exposing
the usb keyboard as the main keyboard and turning it on by default, some
devices do not desire this (e.g. tablets - where it makes the onscreen
keyboard disappear).

This change ensures that when running FAFT, no matter what the servod
configured main keyboard is (usb, or internal) it will be turned on.

BUG=b:131354266
TEST=test_that --autotest_dir . $SARIEN_IP firmware_FAFTSetup
(keyboard is turned on properly, even if it was off on servod
invocation)

Change-Id: I39e9b3ec27022290a2fef6d1af3ede87897fefa0
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1588507
Tested-by: Chris Zhou <chris_zhou@compal.corp-partner.google.com>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>